Refactor sidebar state management and storage utilities
- Moved functions for reading and saving expanded state before browsing to a new utility file, `sidebarBrowserStorage.ts`.
- Updated `SidebarBrowser` and `SidebarNavigation` components to utilize the new storage functions.
- Improved the `useImagePalette` hook to simplify state management for image palette extraction.
- Enhanced `Page` component to manage browse mode transitions more effectively.
