+++ frontend/src/hooks/useImagePalette.ts
@@ -2,24 +2,22 @@ import { useEffect, useState } from 'react';
 import { extractImagePalette, type ImagePalette } from '@/utils/extractImagePalette';
 
 export function useImagePalette(imageUrl: string | undefined, enabled = true) {
+    const active = enabled && !!imageUrl;
     const [palette, setPalette] = useState<ImagePalette | null>(null);
 
     useEffect(() => {
-        if (!enabled || !imageUrl) {
-            setPalette(null);
-            return;
-        }
+        if (!active) return;
 
         let cancelled = false;
 
-        extractImagePalette(imageUrl).then((result) => {
+        extractImagePalette(imageUrl!).then((result) => {
             if (!cancelled) setPalette(result);
         });
 
         return () => {
             cancelled = true;
         };
-    }, [enabled, imageUrl]);
+    }, [active, imageUrl]);
 
-    return palette;
+    return active ? palette : null;
 }
